renamed Utilities to Vector2 for better accuracy
