+++ src/CLGEngine/CORE/Rect.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Utility.h"
+#include "Vector2.h"
 
 struct Rect{
     CLGEngine::CORE::Vector2<float>
