+++ src/CLGEngine/Entity.h
@@ -5,7 +5,7 @@
 #include <map>
 #include <forward_list>
 #include "CORE/ISubject.h"
-#include "CORE/Utility.h"
+#include "CORE/Vector2.h"
 #include "CORE/Rect.h"
