+++ src/CLGEngine/Graphics/Render.h
@@ -3,7 +3,7 @@
 #include <vector>
 #include <Windows.h>
 #include "../CORE/Component.h"
-#include "../CORE/Utility.h"
+#include "../CORE/Vector2.h"
 
 // We can store all created instances to update every cycle
 namespace Graphics {
